requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
async-timeout=4.0.2=pypi_0
beautifulsoup4=4.11.1=pypi_0
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.6.15=py39h06a4308_0
charset-normalizer=2.1.0=pypi_0
deprecated=1.2.13=pypi_0
idna=3.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
ncurses=6.3=h5eee18b_3
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pip=22.1.2=py39h06a4308_0
pyparsing=3.0.9=pypi_0
python=3.9.12=h12debd9_1
python-dotenv=0.20.0=pypi_0
readline=8.1.2=h7f8727e_1
redis=4.3.4=pypi_0
requests=2.28.1=pypi_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spotipy=2.20.0=pypi_0
sqlite=3.38.5=hc218d9a_0
tk=8.6.12=h1ccaba5_0
tzdata=2022a=hda174b7_0
urllib3=1.26.10=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
